Build internal level editor (WIP, doesn't actually run)

+++ Descent3/LoadLevel.cpp
@@ -3470,7 +3470,6 @@ void ReadTerrainChunks(CFILE *fp, int version) {
   UpdateTerrainLightmaps();
 
 #if (defined(EDITOR) || defined(NEWEDITOR))
-
   memset(TerrainSelected, 0, TERRAIN_WIDTH * TERRAIN_DEPTH);
   Num_terrain_selected = 0;
 #endif
@@ -3923,7 +3922,7 @@ int LoadLevel(char *filename, void (*cb_fn)(const char *, int, int)) {
           uint8_t buf[1000];
           int n = chunk_size, r;
           while (n) {
-            r = cf_ReadBytes(buf, std::min(n, sizeof(buf)), ifile);
+            r = cf_ReadBytes(buf, std::min(n, static_cast<int>(sizeof(buf))), ifile);
             cf_WriteBytes(buf, r, ofile);
             n -= r;
           }
